+++ MdeModulePkg/Library/SmmCorePerformanceLib/SmmCorePerformanceLibInternal.h
@@ -5,6 +5,7 @@
   library instance at its constructor.
 
 Copyright (c) 2011 - 2018, Intel Corporation. All rights reserved.<BR>
+Copyright (c) Microsoft Corporation.                                        MU_CHANGE - Standalone MM Perf Support
 SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/
@@ -12,32 +13,135 @@ SPDX-License-Identifier: BSD-2-Clause-Patent
 #ifndef _SMM_CORE_PERFORMANCE_LIB_INTERNAL_H_
 #define _SMM_CORE_PERFORMANCE_LIB_INTERNAL_H_
 
-#include <Guid/Performance.h>
-#include <Guid/PerformanceMeasurement.h>
+// MU_CHANGE [BEGIN] - Standalone MM Perf Support
+#include <Guid/EventGroup.h>
 #include <Guid/ExtendedFirmwarePerformance.h>
 #include <Guid/FirmwarePerformance.h>
+#include <Guid/Performance.h>
+#include <Guid/PerformanceMeasurement.h>
 #include <Guid/ZeroGuid.h>
-#include <Guid/EventGroup.h>
 
-#include <Library/SmmServicesTableLib.h>
-#include <Library/DebugLib.h>
-#include <Library/PerformanceLib.h>
 #include <Library/BaseLib.h>
 #include <Library/BaseMemoryLib.h>
-#include <Library/TimerLib.h>
-#include <Library/PcdLib.h>
-#include <Library/UefiLib.h>
-#include <Library/UefiBootServicesTableLib.h>
+#include <Library/DebugLib.h>
 #include <Library/MemoryAllocationLib.h>
-#include <Library/SynchronizationLib.h>
-#include <Library/SmmMemLib.h>
-#include <Library/ReportStatusCodeLib.h>
-#include <Library/DxeServicesLib.h>
+#include <Library/MmServicesTableLib.h>
+#include <Library/PcdLib.h>
 #include <Library/PeCoffGetEntryPointLib.h>
+#include <Library/PerformanceLib.h>
+#include <Library/SmmMemLib.h>
+#include <Library/SynchronizationLib.h>
+#include <Library/TimerLib.h>
 
-#include <Protocol/SmmBase2.h>
-#include <Protocol/LoadedImage.h>
 #include <Protocol/DevicePathToText.h>
+#include <Protocol/LoadedImage.h>
+
+#define STRING_SIZE              (FPDT_STRING_EVENT_RECORD_NAME_LENGTH * sizeof (CHAR8))
+#define FIRMWARE_RECORD_BUFFER   0x1000
+#define CACHE_HANDLE_GUID_COUNT  0x100
+
+//
+// Library internal function declarations
+//
+
+/**
+  Return a pointer to the loaded image protocol for the given handle.
+
+  @param[in]  Handle      A handle to query for the loaded image protocol.
+
+  @return A pointer to a loaded image protocol instance or null if the handle does not support load image protocol.
+**/
+EFI_LOADED_IMAGE_PROTOCOL *
+GetLoadedImageProtocol (
+  IN EFI_HANDLE  Handle
+  );
+
+/**
+  Get the module name from the PDB file name in the image header.
+
+  @param[in]  ImageBase     The base address of the image.
+  @param[out] NameString    The buffer to store the name string.
+  @param[in]  BufferSize    The size of the buffer in bytes.
+
+  @retval EFI_SUCCESS           The name string is successfully retrieved.
+  @retval EFI_INVALID_PARAMETER A pointer argument provided is null.
+  @retval EFI_NOT_FOUND         The module name was not found.
+
+**/
+EFI_STATUS
+GetModuleNameFromPdbString (
+  IN VOID    *ImageBase,
+  OUT CHAR8  *NameString,
+  IN UINTN   BufferSize
+  );
+
+/**
+  Get the module name from the PDB file name in the image header.
+
+  @param[in]  ModuleGuid    The GUID of the module.
+  @param[out] NameString    The buffer to store the name string.
+  @param[in]  BufferSize    The size of the buffer in bytes.
+
+  @retval EFI_SUCCESS      The name string is successfully retrieved.
+  @retval EFI_NOT_FOUND    The module name was not found.
+
+**/
+EFI_STATUS
+GetNameFromUiSection (
+  IN EFI_GUID  *ModuleGuid,
+  OUT CHAR8    *NameString,
+  IN UINTN     BufferSize
+  );
+
+/**
+  Common initialization code for the MM Core Performance Library.
+
+  @retval     EFI_SUCCESS           The MM Core Performance Library was initialized successfully.
+  @retval     EFI_OUT_OF_RESOURCES  There are not enough resources to initialize the MM Core Performance Library.
+  @retval     Others                The MM Core Performance Library was not initialized successfully.
+ **/
+EFI_STATUS
+InitializeMmCorePerformanceLibCommon (
+  VOID
+  );
+
+/**
+  A library internal MM-instance specific implementation to check if a buffer outside MM is valid.
+
+  This function is provided so Standalone MM and Traditional MM may use a different implementation
+  of data buffer check logic.
+
+  @param[in] Buffer  The buffer start address to be checked.
+  @param[in] Length  The buffer length to be checked.
+
+  @retval TRUE  This buffer is valid per processor architecture.
+  @retval FALSE This buffer is not valid per processor architecture.
+**/
+BOOLEAN
+IsBufferOutsideMmValidInternal (
+  IN EFI_PHYSICAL_ADDRESS  Buffer,
+  IN UINT64                Length
+  );
+
+/**
+  A library internal MM-instance specific implementation to check if a comm buffer is valid.
+
+  This function is provided so Standalone MM and Traditional MM may use a different implementation
+  of comm buffer check logic.
+
+  @param[in] Buffer  The buffer start address to be checked.
+  @param[in] Length  The buffer length to be checked.
+
+  @retval TRUE  This communicate buffer is valid per processor architecture.
+  @retval FALSE This communicate buffer is not valid per processor architecture.
+**/
+BOOLEAN
+IsCommBufferValidInternal (
+  IN EFI_PHYSICAL_ADDRESS  Buffer,
+  IN UINT64                Length
+  );
+
+// MU_CHANGE [END] - Standalone MM Perf Support
 
 //
 // Interface declarations for SMM PerformanceMeasurement Protocol.
